Add new event for marking annotation, create event.py
